refactor to move BackendOptions and BackendHostConfiguration classes into config folder.
