+++ src/SimpleL7Proxy/Program.cs
@@ -11,6 +11,7 @@ using Microsoft.Extensions.Options;
 using Azure.Messaging.ServiceBus;
 
 using SimpleL7Proxy.Backend;
+using SimpleL7Proxy.Config;
 using SimpleL7Proxy.Events;
 using SimpleL7Proxy.Proxy;
 using SimpleL7Proxy.Queue;
